feat(dashboard): matrix card + wiring polish, Add Playbook drawer
Converts the Add Playbook modal into a right-anchored slide-in drawer
(scrollable body, footer with bottom-left owners + Cancel/Save) and
folds in a batch of matrix surface polish.

- Add Playbook: modal → drawer rename, callers updated in process-matrix
  and template-editor-screen; tests and testids carried over to
  `add-playbook-drawer.test.tsx`.
- PlaybookOutputPicker: dropdown auto-flips upward when there isn't
  room below; removed the "PLAYBOOK OUTPUTS" header (dialog keeps its
  accessible name via aria-label).
- Task card: removed the top-right unmet-input arrow glyph; clamped
  `.tc-title` to exactly two lines so wrapped text never leaks below
  the ellipsis after the `.tc-top` wrapper was dropped.
- Matrix mini cell: removed the yellow output-aggregate dot from
  collapsed avatars; dropped the now-unused `aggregateIoStatus` helper
  and `mx-mini-cell-io-dot` CSS.
- Wiring overlay: same-row connector bows upward with a softer arc;
  endpoint anchor dots are suppressed on the side that lands on a mini
  avatar; added a `collapseKey` driven rAF re-measure loop so wires
  follow the 0.22s row/column collapse animation instead of snapping
  on the next hover.
