+++ products/dashboard/app/styles/components/process-matrix.css
@@ -1242,26 +1242,3 @@
   outline-offset: -2px;
 }
 
-/* Aggregate-IO dot anchored to the lower-right of the mini avatar.
- * Replaces the full pip rail when a task lives in a collapsed column /
- * row and the avatar is too small to host individual pips. */
-.mx-mini-cell-io-dot {
-  position: absolute;
-  right: -1px;
-  bottom: -1px;
-  width: 6px;
-  height: 6px;
-  border-radius: 50%;
-  border: 1px solid var(--bg);
-  background: var(--t3);
-  pointer-events: none;
-}
-.mx-mini-cell-io-dot[data-io-status="ok"] {
-  background: var(--ok);
-}
-.mx-mini-cell-io-dot[data-io-status="err"] {
-  background: var(--err);
-}
-.mx-mini-cell-io-dot[data-io-status="pending"] {
-  background: var(--pill-pending-d);
-}
