+++ products/dashboard/app/styles/components/task-card.css
@@ -188,11 +188,6 @@
   background: var(--bg-3);
   opacity: 1;
 }
-.tc-top {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
 .tc-title {
   font-size: 11.5px;
   font-weight: 600;
@@ -205,11 +200,13 @@
   overflow: hidden;
   text-overflow: ellipsis;
   word-break: break-word;
-  flex: 1;
-  min-width: 0;
   /* Reserve two lines of vertical space so single-line titles don't pull
    * the status pill upward — keeps the pill on a stable baseline. */
   min-height: calc(1.3em * 2);
+  /* Cap at exactly two lines too, so `flex-direction: column` on the card
+   * can't stretch the box and let `-webkit-line-clamp`'d lines leak below
+   * the ellipsis. */
+  max-height: calc(1.3em * 2);
 }
 /* Output pip rail (PR 4 / AEL-62). Sits between the title and the status
  * row. One pip per declared `playbook_outputs` row, sorted by position. The
@@ -273,13 +270,6 @@
 .group\/tc-pip:focus-within .tc-pip-tooltip {
   opacity: 1;
 }
-.tc-unmet-input-glyph {
-  color: var(--t3);
-  display: inline-flex;
-  align-items: center;
-  flex-shrink: 0;
-}
-
 .tc-status-row {
   display: flex;
   align-items: center;
